Update sticky.md
Added some content to the overview, tried to clarify the personal wallet being where you store your private key, and noting that wallets should be encrypted when possible. Also re-ordered the desktop wallet links, they were in alphabetic order, and probably (for newbies) should be in order of ease of installation/use.